Refactoring for working with lists of models
